+++ spring-cloud-starter-tencent-polaris-config/src/main/java/com/tencent/cloud/polaris/config/adapter/PolarisConfigFilePuller.java
@@ -112,7 +112,10 @@ public final class PolarisConfigFilePuller {
 	 */
 	public void initCustomPolarisConfigFile(CompositePropertySource compositePropertySource,
 			ConfigFileGroup configFileGroup) {
-		String namespace = polarisContextProperties.getNamespace();
+		String groupNamespace = configFileGroup.getNamespace();
+		if (!StringUtils.hasText(groupNamespace)) {
+			groupNamespace = polarisContextProperties.getNamespace();
+		}
 		String group = configFileGroup.getName();
 		if (!StringUtils.hasText(group)) {
 			throw new IllegalArgumentException("polaris config group name cannot be empty.");
@@ -122,24 +125,23 @@ public final class PolarisConfigFilePuller {
 			return;
 		}
 		for (String fileName : files) {
-			PolarisPropertySource polarisPropertySource = loadPolarisPropertySource(namespace, group, fileName);
+			PolarisPropertySource polarisPropertySource = loadPolarisPropertySource(groupNamespace, group, fileName);
 			compositePropertySource.addPropertySource(polarisPropertySource);
 			PolarisPropertySourceManager.addPropertySource(polarisPropertySource);
 			LOGGER.info(
 					"[SCT Config] Load and inject polaris config file success. namespace = {}, group = {}, fileName = {}",
-					namespace, group, fileName);
+					groupNamespace, group, fileName);
 		}
 	}
 
 	private PolarisPropertySource loadPolarisPropertySource(String namespace, String group, String fileName) {
 		ConfigKVFile configKVFile;
-		// unknown extension is resolved as properties file
-		if (ConfigFileFormat.isPropertyFile(fileName)
-				|| ConfigFileFormat.isUnknownFile(fileName)) {
-			configKVFile = configFileService.getConfigPropertiesFile(namespace, group, fileName);
-		}
-		else if (ConfigFileFormat.isYamlFile(fileName)) {
+		// unknown extension is resolved as yaml file
+		if (ConfigFileFormat.isYamlFile(fileName) || ConfigFileFormat.isUnknownFile(fileName)) {
 			configKVFile = configFileService.getConfigYamlFile(namespace, group, fileName);
+		}
+		else if (ConfigFileFormat.isPropertyFile(fileName)) {
+			configKVFile = configFileService.getConfigPropertiesFile(namespace, group, fileName);
 		}
 		else {
 			LOGGER.warn("[SCT Config] Unsupported config file. namespace = {}, group = {}, fileName = {}", namespace,